+++ test/Controller/Plugin/TestAsset/ListenerStub.php
@@ -1,0 +1,13 @@
+<?php
+namespace ZendTest\Mvc\Controller\Plugin\TestAsset;
+
+class ListenerStub
+{
+    public function __invoke()
+    {
+    }
+
+    public function myCallback()
+    {
+    }
+}